requirements.txt
Django>=3.0,<4.0
psycopg2-binary
python-dotenv==0.15.0
Pillow==8.1.2
django-ckeditor==6.0.0